+++ MMM-DynamicWeather.ts
@@ -96,7 +96,9 @@ Module.register("MMM-DynamicWeather", {
     effectDelay: 60000,
     realisticClouds: false,
     hideSnow: false,
+    hideSnowman: true,
     hideRain: false,
+    hideFlower: true,
     hideClouds: false,
     hideFog: false,
     hideLightning: false,
@@ -343,12 +345,18 @@ Module.register("MMM-DynamicWeather", {
       //Codes from https://openweathermap.org/weather-conditions
       if (this.weatherCode >= 600 && this.weatherCode <= 622 && !this.config.hideSnow) {
         this.showCustomEffect(wrapper, this.snowEffect);
+        if (this.config.hideSnowman === false) {
+          this.buildSnowman(wrapper);
+        }
         if (this.weatherCode >= 611 && this.weatherCode <= 622 && !this.config.hideRain) {
           //snow/rain mix
           this.makeItRain(wrapper);
         }
       } else if (this.weatherCode >= 200 && this.weatherCode <= 531 && !this.config.hideRain) {
         this.makeItRain(wrapper);
+        if (this.config.hideFlower === false) {
+          this.buildFlower(wrapper);
+        }
         if (this.weatherCode >= 200 && this.weatherCode <= 232 && !this.config.hideLightning) {
           this.makeItLightning(wrapper);
         }
@@ -417,14 +425,14 @@ Module.register("MMM-DynamicWeather", {
         }
         case "left-right": {
           flake.className = "flake-left-right";
-          flake.style.left = (-75 - (size * 2)) + "px";
+          flake.style.left = -75 - size * 2 + "px";
           flake.style.top = Math.random() * 100 - 10 + "%";
           flake.style.animationName = "flake-jiggle-left-right";
           break;
         }
         case "right-left": {
           flake.className = "flake-right-left";
-          flake.style.right = (75 + (size * 2)) + "px";
+          flake.style.right = 75 + size * 2 + "px";
           flake.style.top = Math.random() * 100 - 10 + "%";
           flake.style.animationName = "flake-jiggle-right-left";
           break;
@@ -437,9 +445,8 @@ Module.register("MMM-DynamicWeather", {
         }
       }
 
-      //let speed = 95;
-      let max = effect.getSpeedMax(); //(100 - speed) < 1 ? 1 : 100 - speed;
-      let min = effect.getSpeedMin(); //(50 - speed) < 0 ? 0 : 50 - speed;
+      let max = effect.getSpeedMax();
+      let min = effect.getSpeedMin();
 
       jiggle = document.createElement("div");
       jiggle.style.animationDelay = Math.random() * max + "s";
@@ -464,6 +471,16 @@ Module.register("MMM-DynamicWeather", {
 
       wrapper.appendChild(flake);
     }
+  },
+
+  buildSnowman: function (wrapper) {
+    this.doShowEffects = false;
+
+    var snowmanImage = document.createElement("div");
+    snowmanImage.classList.add("snowman");
+    snowmanImage.style.animationDuration = this.config.effectDuration - 10000 + "ms"; //subtract for 10s delay
+
+    wrapper.appendChild(snowmanImage);
   },
 
   makeItRain: function (wrapper) {
@@ -505,6 +522,16 @@ Module.register("MMM-DynamicWeather", {
       wrapper.appendChild(backDrop);
       wrapper.appendChild(frontDrop);
     }
+  },
+
+  buildFlower: function (wrapper) {
+    this.doShowEffects = false;
+
+    var flowerImage = document.createElement("div");
+    flowerImage.classList.add("flower");
+    flowerImage.style.animationDuration = this.config.effectDuration - 10000 + "ms"; //subtract for 10s delay
+
+    wrapper.appendChild(flowerImage);
   },
 
   makeItLightning: function (wrapper) {